Implement feature-specific model usage (#84)
As a user, I would like certain features to use the model that is most effective for the purpose.

**Tasks**

- [x] Make feature-specific model the highest priority
- [x] Indicate the model in the promptContent component
- [x] Test
